requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
anyio==4.1.0
ariadne==0.21
async-timeout==4.0.3
attrs==23.1.0
click==8.1.7
exceptiongroup==1.2.0
frozenlist==1.4.0
graphql-core==3.2.3
h11==0.14.0
idna==3.6
multidict==6.0.4
python-dotenv==1.0.0
sniffio==1.3.0
starlette==0.33.0
typing_extensions==4.8.0
uvicorn==0.24.0.post1
yarl==1.9.3
